pyproject.toml
[project]
name = "lobsterlogreporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyside6-essentials>=6.10.1",
]
